refactor(weekly): decompose WeeklyAggregator into Facade + 3 Mixins
Extract 14 private methods from WeeklyAggregator (904 lines) into 3 mixins:
- DataMixin (6 methods): API and filesystem data fetching
- ProcessingMixin (6 methods): data transformation and aggregation
- AnalysisMixin (2 methods): insights extraction and recommendations

Facade retains __init__, collect_raw_data, process_data, format_output.
MRO: WeeklyAggregator → DataMixin → ProcessingMixin → AnalysisMixin → DataAggregator.
Pattern aligned with analyzers/monthly/ decomposition (PR #177).
