+++ magma_cycling/analyzers/weekly/__init__.py
@@ -1,0 +1,1 @@
+"""WeeklyAggregator mixin modules for god class decomposition."""
