+++ magma_cycling/analyzers/weekly_aggregator.py
@@ -74,12 +74,14 @@ Metadata:
     Version: v2
 """
 
-import json
 import logging
 from datetime import date, timedelta
 from pathlib import Path
 from typing import Any
 
+from magma_cycling.analyzers.weekly.analysis import AnalysisMixin
+from magma_cycling.analyzers.weekly.data import DataMixin
+from magma_cycling.analyzers.weekly.processing import ProcessingMixin
 from magma_cycling.api.intervals_client import IntervalsClient
 from magma_cycling.config import create_intervals_client
 from magma_cycling.core.data_aggregator import DataAggregator
@@ -87,7 +89,7 @@ from magma_cycling.core.data_aggregator import DataAggregator
 logger = logging.getLogger(__name__)
 
 
-class WeeklyAggregator(DataAggregator):
+class WeeklyAggregator(DataMixin, ProcessingMixin, AnalysisMixin, DataAggregator):
     """
     Agrégateur hebdomadaire pour analyse complète semaine.
 
@@ -309,596 +311,3 @@ class WeeklyAggregator(DataAggregator):
             output.append(f"- **TSB :** {metrics.get('tsb', 0):.1f}")
 
         return "\n".join(output)
-
-    # ==================== MÉTHODES PRIVÉES ====================
-
-    def _fetch_weekly_activities(self) -> list[dict[str, Any]]:
-        """
-        Fetch activités semaine depuis Intervals.icu avec détails complets.
-
-        Enrichit chaque activité avec données détaillées (TSS, IF, NP) via
-        appel get_activity() individuel, car get_activities() ne retourne
-        que les champs basiques.
-
-        Returns:
-            Liste activités enrichies avec training_load, if, normalized_power.
-        """
-        if not self.api:
-            logger.warning("No API available, returning empty activities")
-            return []
-
-        start_str = self.start_date.isoformat()
-        end_str = self.end_date.isoformat()
-
-        # Fetch liste activités (données basiques)
-        activities_basic = self.api.get_activities(oldest=start_str, newest=end_str)
-
-        logger.info(f"Found {len(activities_basic)} activities, fetching detailed data...")
-
-        # Enrichir chaque activité avec détails complets
-        activities_detailed = []
-        for activity in activities_basic:
-            activity_id = activity.get("id")
-            if not activity_id:
-                logger.warning(f"Activity without ID: {activity.get('name', 'Unknown')}")
-                activities_detailed.append(activity)
-                continue
-
-            try:
-                # Fetch détails complets (inclut TSS, IF, NP)
-                detailed = self.api.get_activity(activity_id)
-                activities_detailed.append(detailed)
-                logger.debug(
-                    f"Enriched activity {activity_id}: TSS={detailed.get('training_load', 0)}, IF={detailed.get('if', 0)}"
-                )
-            except Exception as e:
-                logger.warning(f"Failed to fetch details for activity {activity_id}: {e}")
-                # Fallback to basic data
-                activities_detailed.append(activity)
-
-        # Trier par date
-        activities_detailed.sort(key=lambda x: x.get("start_date_local", ""))
-
-        logger.info(f"Successfully enriched {len(activities_detailed)} activities with TSS/IF data")
-        return activities_detailed
-
-    def _fetch_daily_metrics(self) -> list[dict[str, Any]]:
-        """Fetch métriques quotidiennes (CTL/ATL/TSB)."""
-        if not self.api:
-            return []
-
-        metrics = []
-        current_date = self.start_date
-
-        while current_date <= self.end_date:
-            try:
-                date_str = current_date.isoformat()
-                wellness_data = self.api.get_wellness(oldest=date_str, newest=date_str)
-
-                # API returns list - convert to dict keyed by date
-                if isinstance(wellness_data, list) and len(wellness_data) > 0:
-                    wellness = wellness_data[0]  # Take first element
-                elif isinstance(wellness_data, dict):
-                    wellness = wellness_data
-                else:
-                    wellness = None
-
-                if wellness:
-                    from magma_cycling.utils.metrics import (
-                        extract_wellness_metrics,
-                    )
-
-                    wellness_metrics = extract_wellness_metrics(wellness)
-                    ramp_rate = wellness.get("ramp_rate")
-
-                    metrics.append(
-                        {
-                            "date": date_str,
-                            "ctl": wellness_metrics["ctl"],
-                            "atl": wellness_metrics["atl"],
-                            "tsb": wellness_metrics["tsb"],
-                            "ramp_rate": ramp_rate if ramp_rate is not None else 0,
-                        }
-                    )
-            except Exception as e:
-                logger.warning(f"No metrics for {current_date}: {e}")
-
-            current_date += timedelta(days=1)
-
-        return metrics
-
-    def _load_weekly_feedback(self) -> dict[str, Any]:
-        """Load feedback athlète pour la semaine."""
-        feedback_dir = self.data_dir / "feedback"
-
-        if not feedback_dir.exists():
-            return {}
-
-        feedback = {}
-        for feedback_file in feedback_dir.glob("*.json"):
-            try:
-                with open(feedback_file, encoding="utf-8") as f:
-                    data = json.load(f)
-                    activity_id = feedback_file.stem
-                    feedback[activity_id] = data
-            except Exception as e:
-                logger.warning(f"Failed to load {feedback_file}: {e}")
-
-        return feedback
-
-    def _fetch_wellness_data(self) -> dict[str, Any]:
-        """Fetch données wellness (sommeil, poids, HRV)."""
-        if not self.api:
-            return {}
-
-        wellness = {}
-        current_date = self.start_date
-
-        while current_date <= self.end_date:
-            try:
-                date_str = current_date.isoformat()
-                wellness_data = self.api.get_wellness(oldest=date_str, newest=date_str)
-
-                # API returns list - convert to dict keyed by date
-                if isinstance(wellness_data, list) and len(wellness_data) > 0:
-                    data = wellness_data[0]  # Take first element
-                elif isinstance(wellness_data, dict):
-                    data = wellness_data
-                else:
-                    data = None
-
-                if data:
-                    wellness[date_str] = {
-                        "sleep_quality": data.get("sleepQuality", 0),
-                        "sleep_hours": (
-                            data.get("sleepSecs", 0) / 3600 if data.get("sleepSecs") else 0
-                        ),
-                        "weight": data.get("weight", 0),
-                        "hrv": data.get("hrvSDNN", 0),
-                        "resting_hr": data.get("restingHR", 0),
-                    }
-            except Exception as e:
-                logger.warning(f"No wellness for {current_date}: {e}")
-
-            current_date += timedelta(days=1)
-
-        return wellness
-
-    def _fetch_planned_workouts(self) -> list[dict[str, Any]]:
-        """Fetch workouts planifiés pour compliance check."""
-        if not self.api:
-            return []
-
-        start_str = self.start_date.isoformat()
-        end_str = self.end_date.isoformat()
-
-        try:
-            planned = self.api.get_events(oldest=start_str, newest=end_str)
-            # Filter for WORKOUT category
-            if planned:
-                planned = [e for e in planned if e.get("category") == "WORKOUT"]
-            return planned
-        except Exception as e:
-            logger.warning(f"Failed to fetch planned workouts: {e}")
-            return []
-
-    def _compute_weekly_summary(self, activities: list[dict[str, Any]]) -> dict[str, Any]:
-        """
-        Calculate summary hebdomadaire.
-
-        Note: Utilise champs Intervals.icu avec préfixe 'icu_':
-        - icu_training_load (TSS)
-        - icu_intensity (IF en %, nécessite normalisation).
-        """
-        # Defensive: filter out None values before summing
-
-        summary = {
-            "total_sessions": len(activities),
-            "total_tss": sum((a.get("icu_training_load") or 0) for a in activities),
-            "total_duration": sum((a.get("moving_time") or 0) for a in activities),
-            "avg_tss": 0,
-            "avg_if": 0,
-            "total_distance": sum((a.get("distance") or 0) for a in activities),
-            "avg_pedal_balance": None,
-            "pedal_balance_imbalance": False,
-        }
-
-        if activities:
-            summary["avg_tss"] = summary["total_tss"] / len(activities)
-
-            # IF moyen (normaliser depuis pourcentage) - Defensive: handle None
-            ifs = [
-                (a.get("icu_intensity") or 0) / 100
-                for a in activities
-                if a.get("icu_intensity") and a.get("icu_intensity") > 0  # type: ignore[operator]  # Safe: checked for None in first condition
-            ]
-            if ifs:
-                summary["avg_if"] = sum(ifs) / len(ifs)
-
-            # Pedal balance moyen
-            balances = [
-                a.get("avg_lr_balance") for a in activities if a.get("avg_lr_balance") is not None
-            ]
-            if balances:
-                avg_balance = sum(balances) / len(balances)
-                summary["avg_pedal_balance"] = avg_balance
-                # Déséquilibre si >52% ou <48% (tolérance 2%)
-                summary["pedal_balance_imbalance"] = avg_balance > 52.0 or avg_balance < 48.0
-
-        # Métriques finales (dernière journée)
-        if activities:
-            last_activity = activities[-1]
-
-            # Extract wellness metrics from last activity
-            from magma_cycling.utils.metrics import extract_wellness_metrics
-
-            final_metrics = extract_wellness_metrics(last_activity)
-            summary["final_metrics"] = final_metrics
-
-        return summary
-
-    def _extract_gear_metrics(self, activity_id: str) -> dict[str, Any] | None:
-        """
-        Extract gear metrics (Di2) from activity streams.
-
-        Extrait métriques de changements de vitesse depuis les streams.
-
-        Args:
-            activity_id: Activity ID
-
-        Returns:
-            Dict avec métriques gear ou None si pas disponibles:
-            - shifts: Nombre total de changements
-            - front_shifts: Changements plateau avant
-            - rear_shifts: Changements pignon arrière
-            - avg_gear_ratio: Ratio moyen
-            - gear_ratio_distribution: Distribution des ratios utilisés
-        """
-        if not self.api:
-            return None
-
-        try:
-            # Fetch streams
-            streams = self.api.get_activity_streams(activity_id)
-
-            # Find gear streams
-            front_gear_stream = next((s for s in streams if s.get("type") == "FrontGear"), None)
-            rear_gear_stream = next((s for s in streams if s.get("type") == "RearGear"), None)
-            gear_ratio_stream = next((s for s in streams if s.get("type") == "GearRatio"), None)
-
-            if not (front_gear_stream and rear_gear_stream):
-                return None  # Pas de données gear disponibles
-
-            front_data = front_gear_stream.get("data", [])
-            rear_data = rear_gear_stream.get("data", [])
-            ratio_data = gear_ratio_stream.get("data", []) if gear_ratio_stream else []
-
-            # Remove None values and calculate metrics
-            front_clean = [f for f in front_data if f is not None]
-            rear_clean = [r for r in rear_data if r is not None]
-            ratio_clean = [r for r in ratio_data if r is not None]
-
-            if not front_clean or not rear_clean:
-                return None
-
-            # Count shifts (changes in gear values)
-            front_shifts = sum(
-                1 for i in range(1, len(front_clean)) if front_clean[i] != front_clean[i - 1]
-            )
-            rear_shifts = sum(
-                1 for i in range(1, len(rear_clean)) if rear_clean[i] != rear_clean[i - 1]
-            )
-
-            total_shifts = front_shifts + rear_shifts
-
-            # Calculate average gear ratio
-            avg_ratio = sum(ratio_clean) / len(ratio_clean) if ratio_clean else None
-
-            # Build gear ratio distribution (top 5 most used)
-            from collections import Counter
-
-            ratio_counts = Counter(round(r, 2) for r in ratio_clean if r)
-            top_ratios = dict(ratio_counts.most_common(5))
-
-            return {
-                "shifts": total_shifts,
-                "front_shifts": front_shifts,
-                "rear_shifts": rear_shifts,
-                "avg_gear_ratio": round(avg_ratio, 2) if avg_ratio else None,
-                "gear_ratio_distribution": top_ratios,
-            }
-
-        except Exception as e:
-            logger.warning(f"Error extracting gear metrics for {activity_id}: {e}")
-            return None
-
-    def _process_workouts_detailed(
-        self, activities: list[dict[str, Any]], feedback: dict[str, Any]
-    ) -> list[dict[str, Any]]:
-        """
-        Process workouts avec détails pour workout_history.
-
-        Map champs Intervals.icu (icu_ prefix) vers format interne:
-        - icu_training_load → tss
-        - icu_intensity (%) → if (normalisé 0.0-1.0)
-        - icu_weighted_avg_watts → normalized_power
-        - icu_average_watts → average_power
-        - avg_lr_balance → pedal_balance (%).
-        """
-        workouts = []
-
-        for i, activity in enumerate(activities, 1):
-            activity_id = str(activity.get("id", ""))
-
-            # Normaliser IF depuis pourcentage (66.36% → 0.66)
-            icu_intensity = activity.get("icu_intensity", 0)
-            if_normalized = icu_intensity / 100 if icu_intensity > 10 else icu_intensity
-
-            workout = {
-                "session_number": i,
-                "date": activity.get("start_date_local", ""),
-                "activity_id": activity_id,
-                "name": activity.get("name", "Unknown"),
-                "type": activity.get("type", "Ride"),
-                "duration": activity.get("moving_time", 0),
-                "tss": activity.get("icu_training_load", 0),
-                "if": if_normalized,
-                "normalized_power": activity.get("icu_weighted_avg_watts", 0),
-                "average_power": activity.get("icu_average_watts", 0),
-                "average_hr": activity.get("average_hr", 0),
-                "max_hr": activity.get("max_hr", 0),
-                "pedal_balance": activity.get("avg_lr_balance"),  # Left/Right balance (%)
-            }
-
-            # Ajouter feedback si disponible
-            if activity_id in feedback:
-                workout["feedback"] = feedback[activity_id]
-
-            # Ajouter métriques gear (Di2) si activité outdoor
-            is_outdoor = activity.get("trainer") is False or activity.get("type") == "Ride"
-            if is_outdoor and activity_id:
-                try:
-                    gear_metrics = self._extract_gear_metrics(activity_id)
-                    if gear_metrics:
-                        workout["gear_metrics"] = gear_metrics
-                        logger.debug(
-                            f"Extracted gear metrics for {activity_id}: "
-                            f"{gear_metrics.get('shifts', 0)} shifts"
-                        )
-                except Exception as e:
-                    logger.warning(f"Failed to extract gear metrics for {activity_id}: {e}")
-
-            workouts.append(workout)
-
-        return workouts
-
-    def _process_metrics_evolution(self, metrics_daily: list[dict[str, Any]]) -> dict[str, Any]:
-        """Process évolution métriques."""
-        if not metrics_daily:
-            return {}
-
-        evolution = {"daily": metrics_daily, "trends": {}}
-
-        # Calculer tendances - Defensive: handle None values
-        if len(metrics_daily) >= 2:
-            first = metrics_daily[0]
-            last = metrics_daily[-1]
-
-            # Calculate metrics change using centralized utility
-            from magma_cycling.utils.metrics import calculate_metrics_change
-
-            evolution["trends"] = calculate_metrics_change(first, last)  # type: ignore[assignment]  # Complex nested dict type inference
-
-        return evolution
-
-    def _extract_training_learnings(
-        self, workouts: list[dict[str, Any]], feedback: dict[str, Any]
-    ) -> list[str]:
-        """Extract enseignements training (pour AI analysis).
-
-        Args:
-            workouts: Processed workouts with enriched data (tss, if, gear_metrics, etc.)
-            feedback: Session feedback data
-        """
-        learnings = []
-
-        # Patterns répétés - Defensive: handle None values
-        high_tss_days = [w for w in workouts if (w.get("tss") or 0) > 80]
-
-        if high_tss_days:
-            learnings.append(f"{len(high_tss_days)} séances haute charge (TSS >80)")
-
-        # IF élevés - Defensive: handle None values
-        high_if_days = [w for w in workouts if (w.get("if") or 0) > 1.0]
-
-        if high_if_days:
-            learnings.append(f"{len(high_if_days)} séances intensité élevée (IF >1.0)")
-
-        # Pedal balance imbalance - Defensive: handle None
-        balances = [w.get("pedal_balance") for w in workouts if w.get("pedal_balance")]
-        if balances:
-            avg_balance = sum(balances) / len(balances)
-            if avg_balance > 52.0:
-                imbalance_pct = avg_balance - 50.0
-                learnings.append(
-                    f"Déséquilibre pédalage détecté: {avg_balance:.1f}% gauche "
-                    f"(+{imbalance_pct:.1f}% vs équilibre)"
-                )
-            elif avg_balance < 48.0:
-                imbalance_pct = 50.0 - avg_balance
-                learnings.append(
-                    f"Déséquilibre pédalage détecté: {avg_balance:.1f}% gauche "
-                    f"(-{imbalance_pct:.1f}% vs équilibre)"
-                )
-
-        # Gear shift analysis (Di2) - Defensive: handle None
-        outdoor_with_gears = [
-            w for w in workouts if w.get("gear_metrics") and w["gear_metrics"].get("shifts")
-        ]
-        if outdoor_with_gears:
-            total_shifts = sum(w["gear_metrics"]["shifts"] for w in outdoor_with_gears)
-            total_duration_hours = sum(w.get("duration", 0) / 3600 for w in outdoor_with_gears)
-
-            if total_duration_hours > 0:
-                shifts_per_hour = total_shifts / total_duration_hours
-
-                # Patterns de changement de vitesse
-                if shifts_per_hour > 50:
-                    learnings.append(
-                        f"Changements de vitesse fréquents: {total_shifts} shifts "
-                        f"({shifts_per_hour:.0f}/h) - considérer anticipation plus fluide"
-                    )
-                elif shifts_per_hour < 20 and total_shifts > 30:
-                    learnings.append(
-                        f"Bonne gestion des vitesses: {total_shifts} shifts "
-                        f"({shifts_per_hour:.0f}/h) - changements bien anticipés"
-                    )
-
-                # Analyse ratio moyen pour sorties outdoor
-                avg_ratios = [
-                    w["gear_metrics"]["avg_gear_ratio"]
-                    for w in outdoor_with_gears
-                    if w["gear_metrics"].get("avg_gear_ratio")
-                ]
-                if avg_ratios:
-                    overall_avg_ratio = sum(avg_ratios) / len(avg_ratios)
-                    if overall_avg_ratio < 1.5:
-                        learnings.append(
-                            f"Développement faible moyen ({overall_avg_ratio:.2f}) - "
-                            "terrain vallonné ou récupération"
-                        )
-                    elif overall_avg_ratio > 3.0:
-                        learnings.append(
-                            f"Développement élevé moyen ({overall_avg_ratio:.2f}) - "
-                            "sorties plates ou tempo soutenu"
-                        )
-
-        # Feedback patterns - Defensive: handle None
-        low_rpe = [
-            fid for fid, f in feedback.items() if f.get("rpe") is not None and f.get("rpe") <= 3
-        ]
-
-        if low_rpe:
-            learnings.append(f"{len(low_rpe)} séances RPE faible (≤3)")
-
-        return learnings
-
-    def _identify_protocol_changes(
-        self, learnings: list[str], metrics_evolution: dict[str, Any]
-    ) -> list[dict[str, Any]]:
-        """Identify changements protocoles nécessaires."""
-        adaptations = []
-
-        # Check TSB trends - Defensive: handle None
-        trends = metrics_evolution.get("trends", {})
-        tsb_change = trends.get("tsb_change")
-
-        if tsb_change is not None and tsb_change < -10:
-            adaptations.append(
-                {
-                    "type": "recovery",
-                    "reason": f"TSB dropped {tsb_change:.1f} points",
-                    "recommendation": "Add recovery day next week",
-                }
-            )
-
-        # Check pedal balance imbalance in learnings
-        for learning in learnings:
-            if "Déséquilibre pédalage" in learning:
-                adaptations.append(
-                    {
-                        "type": "pedal_balance",
-                        "reason": learning,
-                        "recommendation": "Intégrer exercices unilatéraux (single leg drills) "
-                        "et vérifier position/réglages",
-                    }
-                )
-                break  # Only add once
-
-        return adaptations
-
-    def _compute_compliance(
-        self, activities: list[dict[str, Any]], planned: list[dict[str, Any]]
-    ) -> dict[str, Any]:
-        """Calculate compliance planifié vs exécuté."""
-        compliance = {
-            "planned_count": len(planned),
-            "executed_count": len(activities),
-            "rate": 0,
-            "missed": [],
-            "extra": [],
-        }
-
-        if planned:
-            compliance["rate"] = (len(activities) / len(planned)) * 100
-
-        # Identifier séances manquées (simplification)
-        if len(activities) < len(planned):
-            compliance["missed"] = planned[len(activities) :]
-
-        return compliance
-
-    def _prepare_transition_data(
-        self, summary: dict[str, Any], metrics_evolution: dict[str, Any], learnings: list[str]
-    ) -> dict[str, Any]:
-        """Préparer données transition semaine suivante."""
-        transition: dict[str, Any] = {
-            "current_state": {
-                "total_tss": summary.get("total_tss", 0),
-                "avg_tss": summary.get("avg_tss", 0),
-                "final_tsb": summary.get("final_metrics", {}).get("tsb", 0),
-            },
-            "recommendations": [],
-            "focus_areas": learnings[:3] if learnings else [],
-        }
-
-        # Recommandations basées sur TSB
-        tsb = transition["current_state"]["final_tsb"]
-
-        # Handle None TSB gracefully
-        if tsb is not None:
-            if tsb < -15:
-                transition["recommendations"].append("Recovery week recommended (TSB very low)")
-            elif tsb > 10:
-                transition["recommendations"].append("Ready for intensity increase (TSB positive)")
-
-        return transition
-
-    def _analyze_wellness(self, wellness: dict[str, Any]) -> dict[str, Any]:
-        """Analyze données wellness."""
-        insights = {"sleep_quality_avg": 0, "sleep_hours_avg": 0, "weight_trend": 0, "hrv_avg": 0}
-
-        if not wellness:
-            return insights
-
-        # Moyennes - Defensive: handle None values
-        sleep_qualities = [
-            w["sleep_quality"]
-            for w in wellness.values()
-            if w.get("sleep_quality") is not None and w.get("sleep_quality") > 0
-        ]
-
-        if sleep_qualities:
-            insights["sleep_quality_avg"] = sum(sleep_qualities) / len(sleep_qualities)
-
-        sleep_hours = [
-            w["sleep_hours"]
-            for w in wellness.values()
-            if w.get("sleep_hours") is not None and w.get("sleep_hours") > 0
-        ]
-
-        if sleep_hours:
-            insights["sleep_hours_avg"] = sum(sleep_hours) / len(sleep_hours)
-
-        # Tendance poids - Defensive: handle None values
-        weights = [
-            (date, w["weight"])
-            for date, w in wellness.items()
-            if w.get("weight") is not None and w.get("weight") > 0
-        ]
-
-        if len(weights) >= 2:
-            weights.sort()
-            insights["weight_trend"] = weights[-1][1] - weights[0][1]
-
-        return insights
